+++ docs/.vuepress/theme/layouts/Layout.vue
@@ -1,16 +1,12 @@
 <template>
-  <ParentLayout></ParentLayout>
+  <ParentLayout>
+    <template #sidebar-bottom>
+      <Sponsor />
+    </template>
+  </ParentLayout>
 </template>
 
-<script>
-import ParentLayout from '@vuepress/theme-default/lib/client/layouts/Layout.vue';
+<script setup>
+import ParentLayout from '@vuepress/theme-default/layouts/Layout.vue'
 import Sponsor from '../../components/Sponsor.vue';
-
-export default {
-  name: 'Layout',
-  components: {
-    ParentLayout,
-    Sponsor,
-  }
-}
 </script>