+++ src/utils/surge.ts
@@ -186,8 +186,8 @@ function nodeListMapper(
             'https',
             nodeConfig.hostname,
             nodeConfig.port,
-            nodeConfig.username,
-            nodeConfig.password,
+            nodeConfig.username /* istanbul ignore next */ || '',
+            nodeConfig.password /* istanbul ignore next */ || '',
           ].join(', '),
         ].join(' = '),
       ]
@@ -202,8 +202,8 @@ function nodeListMapper(
             'http',
             nodeConfig.hostname,
             nodeConfig.port,
-            nodeConfig.username,
-            nodeConfig.password,
+            nodeConfig.username /* istanbul ignore next */ || '',
+            nodeConfig.password /* istanbul ignore next */ || '',
           ].join(', '),
         ].join(' = '),
       ]
@@ -511,6 +511,8 @@ function appendCommonConfig(
         'sni',
         'serverCertFingerprintSha256',
         'blockQuic',
+        'portHopping',
+        'portHoppingInterval',
       ],
       {
         keyFormat: 'kebabCase',